+++ features/steps/unity_steps.rb
@@ -10,15 +10,19 @@ When("I wait for the mobile game to start") do
   sleep 3
 end
 
+When('I clear all persistent data') do
+  step('I run the "Clear iOS Data" command')
+end
+
 When('I relaunch the Unity mobile app') do
   Maze.driver.launch_app
   # Wait for a fixed time period
   sleep 3
 end
 
-When("I run the {string} mobile scenario") do |scenario|
-
+def dial_number_for(name)
   lookup = {
+      # Scenarios
       "throw Exception" => 1,
       "Log error" => 2,
       "Native exception" => 3,
@@ -33,36 +37,59 @@ When("I run the {string} mobile scenario") do |scenario|
       "Disable Native Errors" => 12,
       "throw Exception with breadcrumbs" => 13,
       "Start SDK no errors" => 14,
-      "Clear iOS Data" => 15
+      "Discard Error Class" => 15,
+      "Java Background Crash" => 16,
+      "Custom App Type" => 17,
+      "Android Persistence Directory" => 18,
+      "Disabled Release Stage" => 19,
+      "Enabled Release Stage" => 20,
+      "Java Background Crash No Threads" => 21,
+      "iOS Native Error" => 22,
+      "iOS Native Error No Threads" => 23,
+
+
+
+      # Commands
+      "Clear iOS Data" => 90
   }
-  number = lookup[scenario]
-  $logger.debug "#{scenario}' has dial-in code #{number}"
+  number = lookup[name]
+  $logger.debug "Command/scenario '#{name}' has dial-in code #{number}"
 
   step("I dial #{number / 10}")
   sleep 1
   step("I dial #{number % 10}")
   sleep 1
-  step("I press dial")
-  sleep 1
+end
+
+When("I run the {string} command") do |command|
+  dial_number_for command
+  step("I press Run Command")
+end
+
+When("I run the {string} mobile scenario") do |scenario|
+  dial_number_for scenario
+  step("I press Run Scenario")
 end
 
 When("I dial {int}") do |number|
   $logger.debug "Dialling #{number}"
-  # Ensure we tap in the button
-  viewport = Maze.driver.session_capabilities['viewportRect']
-  center = viewport['width'] / 2
-  press_at center, 50 + (number * 100)
+  press_at 40 + (number * 80)
   sleep 1
 end
 
-When("I press dial") do
-  # Ensure we tap in the button
-  viewport = Maze.driver.session_capabilities['viewportRect']
-  center = viewport['width'] / 2
-  press_at center, 1050
+When("I press Run Scenario") do
+  press_at 840
 end
 
-def press_at(x, y)
+When("I press Run Command") do
+  press_at 920
+end
+
+def press_at(y)
+
+  # Ensure we tap in the button
+  viewport = Maze.driver.session_capabilities['viewportRect']
+  x = viewport['width'] / 2
 
   $logger.debug "Press at: #{x},#{y}"
 
@@ -89,7 +116,8 @@ When("I run the game in the {string} state") do |state|
     Maze::Runner.environment['BUGSNAG_APIKEY'] = $api_key
     Maze::Runner.environment['MAZE_ENDPOINT'] = endpoint
 
-    command = "open -W #{Maze.config.app} --args -batchmode -nographics"
+    # Call executable directly rather than use open, which flakes on CI
+    command = "#{Maze.config.app}/Contents/MacOS/Mazerunner --args -batchmode -nographics"
     Maze::Runner.run_command(command)
 
   elsif Maze.config.os == 'windows'
@@ -162,7 +190,7 @@ Then("the first significant stack frame methods and files should match:") do |ex
   stacktrace = Maze::Helper.read_key_path(Maze::Server.errors.current[:body], "events.0.exceptions.0.stacktrace")
   expected_frame_values = expected_values.raw
   expected_index = 0
-  
+
   flunk("The stacktrace is empty") if stacktrace.length == 0
   flunk("The stacktrace is not long enough") if stacktrace.length < expected_frame_values.length