+++ tests/BugsnagUnity.Tests/ConfigurationTests.cs
@@ -16,8 +16,8 @@ namespace BugsnagUnity.Payload.Tests
             Assert.IsTrue(config.AutoDetectErrors);
             Assert.IsTrue(config.AutoTrackSessions);
             Assert.AreEqual("production", config.ReleaseStage);
-            Assert.AreEqual("https://notify.bugsnag.com/", config.Endpoint.ToString());
-            Assert.AreEqual("https://sessions.bugsnag.com/", config.SessionEndpoint.ToString());
+            Assert.AreEqual("https://notify.bugsnag.com/", config.Endpoints.Notify.ToString());
+            Assert.AreEqual("https://sessions.bugsnag.com/", config.Endpoints.Session.ToString());
             Assert.AreEqual("foo", config.ApiKey);
         }
 
@@ -32,6 +32,26 @@ namespace BugsnagUnity.Payload.Tests
             Assert.AreEqual(config.MaximumBreadcrumbs, 25);
             config.MaximumBreadcrumbs = 20;
             Assert.AreEqual(config.MaximumBreadcrumbs, 20);
+        }
+
+        [Test]
+        public void EndpointValidation()
+        {
+            var config = new Configuration("foo");
+
+            Assert.IsTrue(config.Endpoints.IsValid);
+
+            config.Endpoint = new System.Uri("https://www.richIsCool.com/");
+
+            Assert.IsFalse(config.Endpoints.IsValid);
+
+            config.SessionEndpoint = new System.Uri("https://www.richIsSuperCool.com/");
+
+            Assert.IsTrue(config.Endpoints.IsValid);
+
+            config.Endpoint = null;
+
+            Assert.IsFalse(config.Endpoints.IsValid);
         }
     }
 }